Carrinho

+++ app/src/main/java/com/example/projeto_jeitinho_brasileiro/repositorio/receita/Receita.kt
@@ -3,7 +3,7 @@ package com.example.projeto_jeitinho_brasileiro.repositorio.receita
 import com.example.projeto_jeitinho_brasileiro.repositorio.conteudo.Repositorio
 
 data class Receita(
-    val indice: Int = 0,
+    val indice: Long = 0,
     var arquivo: String = "",
     var descricao: String = "",
     var nome: String = "",
